requirements.txt
Flask==2.0.2
folium==0.12.1.post1
geopy==2.2.0
python-dotenv==0.19.2
requests==2.26.0